requirements-tools.txt
slither-analyzer>=0.10.0
hypothesis>=6.92.0
web3>=6.0.0
solc-select>=1.0.4
py-solc-x>=2.0.0
tree-sitter>=0.20.0
tree-sitter-solidity>=0.1.0
graphviz>=0.20.0
matplotlib>=3.7.0
pandas>=2.0.0
pyevmasm>=0.2.3
crytic-compile>=0.3.5
scikit-learn>=1.3.0
numpy>=1.24.0
replicate>=0.20.0
transformers>=4.35.0
torch>=2.1.0
langchain>=0.1.0
langchain-openai>=0.0.2
requests>=2.31.0
aiohttp>=3.9.0
